refactor: don't ever use TS for frontend pls
