+++ frontend/next.config.js
@@ -8,6 +8,10 @@ const nextConfig = {
         poll: 1000,
         aggregateTimeout: 300
       }
+      config.experiments = {
+        ...config.experiments,
+        topLevelAwait: true,
+      }
       return config
     }
   }
