+++ frontend/package.json
@@ -17,14 +17,10 @@
     "react-dom": "^18"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
     "autoprefixer": "^10.4.16",
     "eslint": "^8",
     "eslint-config-next": "13.5.4",
     "postcss": "^8.4.31",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^5"
+    "tailwindcss": "^3.3.3"
   }
 }
