+++ frontend/src/app/layout.jsx
@@ -1,18 +1,17 @@
 import "./globals.css";
-import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 
 import Providers from "./providers";
-import Nav from "@/components/Nav/Nav";
+import Nav from "../components/Nav/Nav";
 
 const inter = Inter({ subsets: ["latin"] })
 
-export const metadata: Metadata = {
+export const metadata = {
   title: "TagMedX",
   description: "TagMedX is an open-source web app built with FastAPI, Next.js, and MySQL, designed for medical image tagging."
 }
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function RootLayout({ children }) {
   return (
     <html lang="en" className="light">
       <body className={inter.className}>
