+++ frontend/src/app/login/page.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import LoginForm from "@/components/LoginForm/LoginForm"
+import LoginForm from "../../components/LoginForm/LoginForm"
 
 export default function LoginPage() {
     return (
