+++ frontend/src/app/signup/page.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import SignupForm from "@/components/SignupForm/SignupForm"
+import SignupForm from "../../components/SignupForm/SignupForm"
 
 export default function SignupPage () {
     return (
