+++ frontend/src/components/LoginForm/LoginFormConsts.js
@@ -1,0 +1,4 @@
+export const defaultLoginData = {
+    email: '',
+    password: ''
+}
