--- frontend/src/components/LoginForm/LoginFormConsts.ts
@@ -1,9 +1,0 @@
-export interface ILoginData {
-    email: string,
-    password: string
-}
-
-export const defaultLoginData: ILoginData = {
-    email: '',
-    password: ''
-}
