+++ frontend/src/components/LoginForm/LoginFormUtils.js
@@ -1,0 +1,3 @@
+export function handleLogin(setSent, data) {
+    setSent(true)
+}
