--- frontend/src/components/LoginForm/LoginFormUtils.ts
@@ -1,5 +1,0 @@
-import { ILoginData } from "./LoginFormConsts";
-
-export function handleLogin(setSent: any, data: ILoginData): void {
-    setSent(true)
-}
