+++ frontend/src/components/NavDesktop/NavDesktopConsts.js
@@ -2,4 +2,4 @@ export const menuItems = [
     { name: "Page 1", url: "page1", color: 'foreground' },
     { name: "Page 2", url: "page2", color: 'foreground' },
     { name: "Page 3", url: "page3", color: 'foreground' }
-] as const;
+];
