+++ frontend/src/components/NavMobile/NavMobileConsts.js
@@ -4,4 +4,4 @@ export const menuItems = [
     { name: "Page 3", url: "page3", color: 'foreground' },
     { name: "Sign Up", url: "signup", color: 'foreground' },
     { name: "Login", url: "login", color: 'primary' }
-] as const;
+];
