+++ frontend/src/components/SignupForm/SignupFormConsts.js
@@ -1,0 +1,6 @@
+export const defaultSignupData = {
+    name: '',
+    surname: '',
+    email: '',
+    password: ''
+};
