--- frontend/src/components/SignupForm/SignupFormConsts.ts
@@ -1,20 +1,0 @@
-export interface ISignupData {
-    name: string,
-    surname: string,
-    email: string,
-    password: string
-}
-
-export const defaultSignupData: ISignupData = {
-    name: '',
-    surname: '',
-    email: '',
-    password: ''
-}
-
-export interface ISignupErrorData {
-    name: boolean,
-    surname: boolean,
-    email: boolean,
-    password: boolean
-}
