+++ frontend/src/consts/NextColor.js
@@ -1,0 +1,8 @@
+export const NextColor = {
+    Default: "default",
+    Primary: "primary",
+    Secondary: "secondary",
+    Success: "success",
+    Warning: "warning",
+    Danger: "danger"
+}
