--- frontend/src/consts/NextColor.ts
@@ -1,8 +1,0 @@
-export enum NextColor {
-    Default = "default",
-    Primary = "primary",
-    Secondary = "secondary",
-    Success = "success",
-    Warning = "warning",
-    Danger = "danger"
-}
