--- frontend/src/types/globals.d.ts
@@ -1,12 +1,0 @@
-export {};
-
-declare global {
-
-    interface IUser {
-        id: number
-        name: string
-        surname: string
-    }
-
-    // type NextColor = "default" | "primary" | "secondary" | "success" | "warning" | "danger";
-}
